Fixed project structure. I restructured the project, now every test is in unitTests.csproj in the same solution as UValueLibrary. (added a reference for UValueLibrary.csproj in the test project) This should remove the file redundancy.
