--- Program.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace uTest
-{
-    class Program
-    {
-        public static void Main(string[] args)
-        {   
-            
-        }
-
-    }
-}
