+++ UValueLibrary/Component.cs
@@ -2,60 +2,43 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace uTest
+namespace validTests
 {
-    class Component
+    public class Component
     {
+        //private readonly List<Layer> layers;
+        public List<Layer> layers;
 
-        public enum ComponentType
+        public ComponentType type;
+
+        public enum ComponentType {wall, floor}
+
+        public Component(List<Layer> layers, ComponentType type)
         {
-            Wall,
-            CeilingFloor,
-            Window,
-            Door
-        }
-
-        private readonly List<Layer> layers;
-        private ComponentType componentType;
-
-
-        public Component(ComponentType componentType, List<Layer> layers)
-        {
-            
-            this.componentType = componentType;
-
-            bool layersCompatible = true;
-            foreach (Layer layer in layers)
-            {
-               if (!layer.CheckCompatibility(this.componentType))
-                {
-                    layersCompatible = false;
-                }
-            }
-            if (!layersCompatible)
-            {
-                throw new ArgumentException("Given layers are incompatible with this component type.");
-            }
-
+            this.layers = layers;
+            this.type = type;
         }
         /// <summary>
         /// Returns a U Value of this component.
         /// </summary>
         /// <param name="Rse">External heat transfer resistance.</param>
-        /// <param name="Rsi">Internal heat trasnfer resistance.</param>
+        /// <param name="Rsi">Internal heat transfer resistance.</param>
         /// <returns></returns>
+        /// 
+
         public double GetUValue(double Rse, double Rsi)
         {
             return 1 / (this.GetRValueSum() + Rse + Rsi);
         }
 
-        private double GetRValueSum()
+        public double GetRValueSum()
         {
             double output = 0;
             foreach (Layer layer in layers)
             {
                 output += layer.GetRValue();
             }
+
             return output;
         }
     }
