+++ uValueProject.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29102.190
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "uTest", "uTest\uTest.csproj", "{EA0CCB3A-D599-4D1A-9068-AB85ECB4D82E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "unitTests", "uTest\unitTests.csproj", "{EA0CCB3A-D599-4D1A-9068-AB85ECB4D82E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UValueLibrary", "UValueLibrary\UValueLibrary.csproj", "{C5FFACC1-AFAC-4FBC-8653-8171CF82DE1F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UValueLibrary", "UValueLibrary\UValueLibrary.csproj", "{C5FFACC1-AFAC-4FBC-8653-8171CF82DE1F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
